pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Kieran Pritchard <kieran.pritchard@ons.gov.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jwt = "^1.3.1"
python-dateutil = "^2.9.0.post0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
